TST: test openly accessible tiles
